feat(build): Copy entire iob_linux submodule in build dir; Fix default.nix.
