--- iob_linux.py
@@ -1,12 +1,0 @@
-# SPDX-FileCopyrightText: 2025 IObundle
-#
-# SPDX-License-Identifier: MIT
-
-
-def setup(py_params_dict):
-    attributes_dict = {
-        "generate_hw": False,
-        "instantiate": False,
-    }
-
-    return attributes_dict
